requirements.txt
aiohttp==3.7.3
alembic==1.14.1
async-timeout==3.0.1
attrs==20.3.0
bcrypt==4.0.1
blinker==1.9.0
chardet==3.0.4
click==8.1.8
colorama==0.4.6
discord.py==1.6.0
dotenv==0.9.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
multidict==5.1.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.6.3
zipp==3.21.0
